refactor(circuit): extract SubsystemTreeMixin from CircuitRoutines [Clean Code: SRP]
Move the seven-method "build/refresh the hierarchical-diagonalization
subsystem tree" cluster (formerly lines 938-1172 of
circuit_internals/routines.py, ~235 lines) into a new sibling mixin:

* ``_check_truncation_indices``
* ``_generate_subsystems``
* ``_get_systems_and_interactions``
* ``_update_existing_subsystems``
* ``_create_new_subsystems``
* ``get_subsystem_index``
* ``_update_interactions``

These methods form a coherent block — they all run during the
"split this Hamiltonian into subsystems and wire them together"
phase of ``_configure``.  Splitting them out lets the rest of
``routines.py`` focus on per-instance concerns (operator
construction, parameter sync) without the tree-construction noise.

CircuitRoutines now subclasses ``SubsystemTreeMixin``; the moved
methods are inherited transparently, so all call sites (in
``Circuit._configure``, ``Subsystem.__init__``, the lifecycle
``update`` path, etc.) continue to resolve them via attribute
lookup.  The new module declares the cross-mixin attributes /
callables it consumes under ``TYPE_CHECKING`` so mypy can resolve
``self.<x>`` references without affecting runtime.

routines.py: 2447 -> 2211 lines.  ~10% smaller, single most
self-contained cluster removed.  The reviewer-recommended full
split (lifecycle + Hamiltonian-assembly mixins) is deferred to a
follow-up session — those clusters cross-reference each other and
the residual core more heavily, so they need characterization
tests in place before extraction.  This Tier 5a extraction is a
no-risk first step.

Full pytest (309 passed, 14 skipped); mypy clean (100 source files,
+1 for the new module); black/isort no diff.
